update .gitignore to include out directory; name change System to Manager; subclass UserManager; some detail changes to Manager (former System)
